requirements.txt
Flask[async]==2.3.3
Flask-Talisman==1.0.0
requests==2.31.0
Flask-Assets==2.0 
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0
watchdog>=3.0.0
Flask-BasicAuth==0.2.0 